Move RichEvent to its own unit
