+++ frontend/app/model/EventPortfolio.scala
@@ -1,6 +1,6 @@
 package model
 
-import model.Eventbrite.RichEvent
+import model.RichEvent.RichEvent
 
 case class EventPortfolio(orderedEvents: Seq[RichEvent], normal: Seq[RichEvent], pastEvents: Option[Seq[RichEvent]]) {
   lazy val heroOpt = orderedEvents.headOption
