+++ frontend/app/views/event/masterclass.scala.html
@@ -5,7 +5,7 @@
     selectedSubTag: String = ""
 )(implicit token: play.filters.csrf.CSRF.Token)
 
-@import model.Eventbrite.MasterclassEvent
+@import model.RichEvent.MasterclassEvent
 
 @main("Masterclasses", htmlClass="events-bg theme--masterclasses", pageInfo=pageInfo) {
