+++ frontend/app/views/event/thankyou.scala.html
@@ -1,4 +1,4 @@
-@(event: model.Eventbrite.RichEvent, order: model.Eventbrite.EBOrder)
+@(event: model.RichEvent.RichEvent, order: model.Eventbrite.EBOrder)
 
 @import configuration.Config
 @import views.support.Dates._
