+++ frontend/app/views/fragments/event/addressSummary.scala.html
@@ -1,4 +1,4 @@
-@(event: model.Eventbrite.RichEvent)
+@(event: model.RichEvent.RichEvent)
 
 <div class="event-location">
     @if(event.venue.name) {
