+++ frontend/app/views/fragments/event/item.scala.html
@@ -1,5 +1,4 @@
-@(event: model.Eventbrite.RichEvent)
-
+@(event: model.RichEvent.RichEvent)
 @import views.support.Dates._
 @import configuration.Config
 @import model.Eventbrite._
