+++ frontend/app/views/fragments/event/itemMetaAvailability.scala.html
@@ -1,4 +1,4 @@
-@(event: model.Eventbrite.RichEvent)
+@(event: model.RichEvent.RichEvent)
 
 @import model.Eventbrite._
