+++ frontend/app/views/fragments/event/itemMetaTitle.scala.html
@@ -1,4 +1,4 @@
-@(event: model.Eventbrite.RichEvent)
+@(event: model.RichEvent.RichEvent)
 
 <h4 class="event-item__title">
     @if(event.isSoldOut){
