+++ frontend/app/views/fragments/event/itemMinimal.scala.html
@@ -1,4 +1,4 @@
-@(event: model.Eventbrite.RichEvent, isCard: Boolean = false)
+@(event: model.RichEvent.RichEvent, isCard: Boolean = false)
 
 <div class="event-item @if(isCard){ event-item--card} else { event-item--minimal}" itemscope itemtype="http://schema.org/Event">
     <div class="event-item__container">
