+++ frontend/app/views/fragments/event/itemSnapshot.scala.html
@@ -1,4 +1,4 @@
-@(event: model.Eventbrite.RichEvent)
+@(event: model.RichEvent.RichEvent)
 
 <div class="event-snapshot" itemscope itemtype="http://schema.org/Event">
     <div class="event-snapshot__media" itemprop="image" content="@event.socialImgUrl">
