+++ frontend/app/views/fragments/event/stats.scala.html
@@ -1,4 +1,4 @@
-@(event: model.Eventbrite.RichEvent)
+@(event: model.RichEvent.RichEvent)
 
 @import views.support.Dates._
