+++ frontend/app/views/fragments/event/ticketCta.scala.html
@@ -1,6 +1,6 @@
-@(event: model.Eventbrite.RichEvent, actionClassSoldOut: String, actionClassBuy: String)
+@(event: model.RichEvent.RichEvent, actionClassSoldOut: String, actionClassBuy: String)
 
-@import model.Eventbrite._
+@import model.RichEvent._
 @import configuration.Config
 
 @termsAndConditions = @{
