+++ frontend/app/views/fragments/event/ticketSales.scala.html
@@ -1,9 +1,9 @@
-@(event: model.Eventbrite.RichEvent, ticket: model.Eventbrite.EBTicketClass)
+@(event: model.RichEvent.RichEvent, ticket: model.Eventbrite.EBTicketClass)
 
 @import views.support.Dates._
 @import com.github.nscala_time.time.Imports._
 @import org.joda.time.Instant
-@import model.Eventbrite._
+@import model.RichEvent._
 @import com.gu.membership.salesforce.Tier
