+++ frontend/app/views/fragments/pricing/priceInfoEvent.scala.html
@@ -1,4 +1,4 @@
-@(event: model.Eventbrite.RichEvent)
+@(event: model.RichEvent.RichEvent)
 
 @for(ticket <- event.generalReleaseTicket) {
     <div class="price-info-inline">
