+++ frontend/app/views/joining/tierChooser.scala.html
@@ -1,4 +1,4 @@
-@(eventOpt: Option[model.Eventbrite.RichEvent], pageInfo: model.PageInfo)(implicit token: play.filters.csrf.CSRF.Token)
+@(eventOpt: Option[model.RichEvent.RichEvent], pageInfo: model.PageInfo)(implicit token: play.filters.csrf.CSRF.Token)
 
 @import com.gu.membership.salesforce.Tier
 @import views.support.Dates._
